[hailtop] ensure delete also deletes single files
I rearranged some fixtures as well.
