+++ hail/python/test/hailtop/inter_cloud/utils.py
@@ -1,0 +1,10 @@
+from typing import Dict
+import secrets
+from hailtop.aiotools.fs import AsyncFS
+
+
+async def fresh_dir(fs: AsyncFS, bases: Dict[str, str], scheme: str):
+    token = secrets.token_hex(16)
+    dir = f'{bases[scheme]}{token}/'
+    await fs.mkdir(dir)
+    return dir
